requirements.txt
torch
torchvision
transformers
pandas
numpy
scikit-learn
pillow
pyyaml
tqdm
